refactor: mv _get_licenses to abstracte extractor
