+++ gimie/sources/abstract.py
@@ -21,6 +21,7 @@ from typing import List, Optional
 
 from rdflib import Graph
 from urllib.parse import urlparse
+from gimie.sources.common.license import _get_license_url, is_license_path
 from gimie.io import Resource
 
 
@@ -83,3 +84,16 @@ class Extractor(ABC):
             url = urlparse(self.url)
             return f"{url.scheme}://{url.netloc}"
         return self.base_url
+
+    def _get_licenses(self) -> list[str]:
+        """Extract a SPDX License URL from a GitHub Repository"""
+        license_files_iterator = filter(
+            lambda p: is_license_path(p.name), self.list_files()
+        )
+        license_files = list(license_files_iterator)
+        license_urls = []
+        for file in license_files:
+            license_url = _get_license_url(file)
+            if license_url:
+                license_urls.append(license_url)
+        return license_urls
