+++ gimie/sources/gitlab.py
@@ -286,21 +286,6 @@ class GitlabExtractor(Extractor):
             email=node.get("publicEmail"),
         )
 
-    def _get_licenses(self) -> list[str]:
-        """Extract a SPDX License URL from a GitLab Repository"""
-        license_files_iterator = filter(
-            lambda p: is_license_path(p.name), self.list_files()
-        )
-        license_files = list(license_files_iterator)
-        license_urls = []
-        for file in license_files:
-            with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-                temp_file.write(file.open().read())
-                license_url = _get_license_url(temp_file.name)
-                if license_urls:
-                    license_urls.append(license_url)
-        return license_urls
-
     def _user_from_rest(self, username: str) -> Person:
         """Given a username, use the REST API to retrieve the Person object."""
